crud-2 motorcycle

+++ migrations/20210929155405-create-motorcycle.js
@@ -16,7 +16,7 @@ module.exports = {
         }
       },
       price: {
-        type: DataTypes.INTEGER(4),
+        type: DataTypes.INTEGER(40),
         allowNull: false,
         validate: {
           notNull: true
